requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
selenium>=4.5.0
pandas>=1.5.0
notion-client>=1.0.0
google-generativeai>=0.5.0
